refactor: combine groups with config and member requests modules
