+++ src/comments/comments.service.ts
@@ -4,7 +4,7 @@ import { FileUpload } from 'graphql-upload-ts';
 import { FindOptionsWhere, In, Repository } from 'typeorm';
 import { DEFAULT_PAGE_SIZE } from '../common/common.constants';
 import { sanitizeText } from '../common/common.utils';
-import { GroupPrivacy } from '../groups/group-configs/group-configs.constants';
+import { GroupPrivacy } from '../groups/groups.constants';
 import { deleteImageFile, saveImage } from '../images/image.utils';
 import { Image } from '../images/models/image.model';
 import { User } from '../users/models/user.model';
