+++ src/database/data-source.ts
@@ -4,8 +4,8 @@ import { Canary } from '../canaries/models/canary.model';
 import { Comment } from '../comments/models/comment.model';
 import { EventAttendee } from '../events/event-attendees/models/event-attendee.model';
 import { Event } from '../events/models/event.model';
-import { GroupConfig } from '../groups/group-configs/models/group-config.model';
-import { GroupMemberRequest } from '../groups/group-member-requests/models/group-member-request.model';
+import { GroupConfig } from '../groups/models/group-config.model';
+import { GroupMemberRequest } from '../groups/models/group-member-request.model';
 import { GroupRolePermission } from '../groups/group-roles/models/group-role-permission.model';
 import { GroupRole } from '../groups/group-roles/models/group-role.model';
 import { Group } from '../groups/models/group.model';
