+++ src/groups/group-configs.resolver.ts
@@ -6,16 +6,16 @@ import {
   ResolveField,
   Resolver,
 } from '@nestjs/graphql';
-import { Dataloaders } from '../../dataloader/dataloader.types';
-import { Group } from '../models/group.model';
-import { UpdateGroupPayload } from '../models/update-group.payload';
-import { GroupConfigsService } from './group-configs.service';
+import { Dataloaders } from '../dataloader/dataloader.types';
+import { GroupsService } from './groups.service';
 import { GroupConfig } from './models/group-config.model';
+import { Group } from './models/group.model';
 import { UpdateGroupConfigInput } from './models/update-group-config.input';
+import { UpdateGroupPayload } from './models/update-group.payload';
 
 @Resolver(() => GroupConfig)
 export class GroupConfigsResolver {
-  constructor(private groupConfigsService: GroupConfigsService) {}
+  constructor(private groupsService: GroupsService) {}
 
   @ResolveField(() => Group)
   async group(
@@ -27,13 +27,13 @@ export class GroupConfigsResolver {
 
   @ResolveField(() => Boolean)
   async isPublic(@Parent() { id }: GroupConfig) {
-    return this.groupConfigsService.getIsPublic(id);
+    return this.groupsService.getIsPublic(id);
   }
 
   @Mutation(() => UpdateGroupPayload)
   async updateGroupConfig(
     @Args('groupConfigData') groupConfigData: UpdateGroupConfigInput,
   ) {
-    return this.groupConfigsService.updateGroupConfig(groupConfigData);
+    return this.groupsService.updateGroupConfig(groupConfigData);
   }
 }
