+++ src/groups/group-member-requests.resolver.ts
@@ -10,25 +10,25 @@ import {
   ResolveField,
   Resolver,
 } from '@nestjs/graphql';
-import { CurrentUser } from '../../auth/decorators/current-user.decorator';
-import { Dataloaders } from '../../dataloader/dataloader.types';
-import { User } from '../../users/models/user.model';
-import { Group } from '../models/group.model';
-import { GroupMemberRequestsService } from './group-member-requests.service';
+import { CurrentUser } from '../auth/decorators/current-user.decorator';
+import { Dataloaders } from '../dataloader/dataloader.types';
+import { User } from '../users/models/user.model';
+import { GroupsService } from './groups.service';
 import { ApproveGroupMemberRequestPayload } from './models/approve-group-member-request.payload';
 import { CreateGroupMemberRequestPayload } from './models/create-group-member-request.payload';
 import { GroupMemberRequest } from './models/group-member-request.model';
+import { Group } from './models/group.model';
 
 @Resolver(() => GroupMemberRequest)
 export class GroupMemberRequestsResolver {
-  constructor(private groupMemberRequestsService: GroupMemberRequestsService) {}
+  constructor(private groupsService: GroupsService) {}
 
   @Query(() => GroupMemberRequest, { nullable: true })
   async groupMemberRequest(
     @Args('groupId', { type: () => Int }) groupId: number,
     @CurrentUser() { id }: User,
   ) {
-    return this.groupMemberRequestsService.getGroupMemberRequest({
+    return this.groupsService.getGroupMemberRequest({
       user: { id },
       groupId,
     });
@@ -55,24 +55,21 @@ export class GroupMemberRequestsResolver {
     @Args('groupId', { type: () => Int }) groupId: number,
     @CurrentUser() { id }: User,
   ) {
-    return this.groupMemberRequestsService.createGroupMemberRequest(
-      groupId,
-      id,
-    );
+    return this.groupsService.createGroupMemberRequest(groupId, id);
   }
 
   @Mutation(() => ApproveGroupMemberRequestPayload)
   async approveGroupMemberRequest(@Args('id', { type: () => Int }) id: number) {
-    return this.groupMemberRequestsService.approveGroupMemberRequest(id);
+    return this.groupsService.approveGroupMemberRequest(id);
   }
 
   @Mutation(() => Boolean)
   async cancelGroupMemberRequest(@Args('id', { type: () => Int }) id: number) {
-    return this.groupMemberRequestsService.cancelGroupMemberRequest(id);
+    return this.groupsService.cancelGroupMemberRequest(id);
   }
 
   @Mutation(() => Boolean)
   async denyGroupMemberRequest(@Args('id', { type: () => Int }) id: number) {
-    return this.groupMemberRequestsService.denyGroupMemberRequest(id);
+    return this.groupsService.denyGroupMemberRequest(id);
   }
 }
