+++ src/groups/groups.module.ts
@@ -3,18 +3,18 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { EventsModule } from '../events/events.module';
 import { Image } from '../images/models/image.model';
 import { UsersModule } from '../users/users.module';
-import { GroupConfigsModule } from './group-configs/group-configs.module';
-import { GroupMemberRequestsModule } from './group-member-requests/group-member-requests.module';
+import { GroupConfigsResolver } from './group-configs.resolver';
+import { GroupMemberRequestsResolver } from './group-member-requests.resolver';
 import { GroupRolesModule } from './group-roles/group-roles.module';
 import { GroupsResolver } from './groups.resolver';
 import { GroupsService } from './groups.service';
+import { GroupConfig } from './models/group-config.model';
+import { GroupMemberRequest } from './models/group-member-request.model';
 import { Group } from './models/group.model';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Group, Image]),
-    forwardRef(() => GroupConfigsModule),
-    forwardRef(() => GroupMemberRequestsModule),
+    TypeOrmModule.forFeature([Group, GroupConfig, Image, GroupMemberRequest]),
 
     // TODO: Remove forwardRef() when circular dependency is fixed
     forwardRef(() => UsersModule),
@@ -22,7 +22,12 @@ import { Group } from './models/group.model';
     EventsModule,
     GroupRolesModule,
   ],
-  providers: [GroupsService, GroupsResolver],
+  providers: [
+    GroupConfigsResolver,
+    GroupMemberRequestsResolver,
+    GroupsResolver,
+    GroupsService,
+  ],
   exports: [GroupsService, TypeOrmModule],
 })
 export class GroupsModule {}
