+++ src/groups/groups.service.ts
@@ -1,5 +1,5 @@
 import { UserInputError } from '@nestjs/apollo';
-import { forwardRef, Inject, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { FileUpload } from 'graphql-upload-ts';
 import { FindOptionsWhere, In, Repository } from 'typeorm';
@@ -14,18 +14,23 @@ import {
 import { Image } from '../images/models/image.model';
 import { Post } from '../posts/models/post.model';
 import { Proposal } from '../proposals/models/proposal.model';
+import { DecisionMakingModel } from '../proposals/proposals.constants';
 import { UsersService } from '../users/users.service';
-import { GroupPrivacy } from './group-configs/group-configs.constants';
-import { GroupConfigsService } from './group-configs/group-configs.service';
-import { GroupMemberRequestsService } from './group-member-requests/group-member-requests.service';
 import { initGroupRolePermissions } from './group-roles/group-role.utils';
 import { GroupRolesService } from './group-roles/group-roles.service';
-import { GroupAdminModel } from './groups.constants';
+import { GroupAdminModel, GroupPrivacy } from './groups.constants';
 import { CreateGroupInput } from './models/create-group.input';
+import { GroupConfig } from './models/group-config.model';
+import {
+  GroupMemberRequest,
+  GroupMemberRequestStatus,
+} from './models/group-member-request.model';
 import { Group } from './models/group.model';
+import { UpdateGroupConfigInput } from './models/update-group-config.input';
 import { UpdateGroupInput } from './models/update-group.input';
 
 type GroupWithMemberCount = Group & { memberCount: number };
+type GroupWithMemberRequestCount = Group & { memberRequestCount: number };
 
 @Injectable()
 export class GroupsService {
@@ -33,13 +38,15 @@ export class GroupsService {
     @InjectRepository(Group)
     private groupRepository: Repository<Group>,
 
+    @InjectRepository(GroupConfig)
+    private groupConfigRepository: Repository<GroupConfig>,
+
+    @InjectRepository(GroupMemberRequest)
+    private groupMemberRequestRepository: Repository<GroupMemberRequest>,
+
     @InjectRepository(Image)
     private imageRepository: Repository<Image>,
 
-    @Inject(forwardRef(() => GroupMemberRequestsService))
-    private memberRequestsService: GroupMemberRequestsService,
-
-    private groupConfigsService: GroupConfigsService,
     private groupRolesService: GroupRolesService,
     private usersService: UsersService,
   ) {}
@@ -129,7 +136,7 @@ export class GroupsService {
   }
 
   async isNoAdminGroup(groupId: number) {
-    const config = await this.groupConfigsService.getGroupConfig({ groupId });
+    const config = await this.getGroupConfig({ groupId });
     return config.adminModel === GroupAdminModel.NoAdmin;
   }
 
@@ -231,7 +238,7 @@ export class GroupsService {
     } else {
       await this.saveDefaultCoverPhoto(group.id);
     }
-    await this.groupConfigsService.initGroupConfig(group.id);
+    await this.initGroupConfig(group.id);
     await this.groupRolesService.initGroupAdminRole(userId, group.id);
 
     return { group };
@@ -317,7 +324,7 @@ export class GroupsService {
   async leaveGroup(id: number, userId: number) {
     const where = { group: { id }, userId };
     await this.deleteGroupMember(id, userId);
-    await this.memberRequestsService.deleteGroupMemberRequest(where);
+    await this.deleteGroupMemberRequest(where);
     return true;
   }
 
@@ -328,5 +335,127 @@ export class GroupsService {
       filename,
       groupId,
     });
+  }
+
+  async getGroupConfig(where: FindOptionsWhere<GroupConfig>) {
+    return this.groupConfigRepository.findOneOrFail({ where });
+  }
+
+  async getIsPublic(id: number) {
+    const groupConfig = await this.getGroupConfig({ id });
+    return groupConfig.privacy === GroupPrivacy.Public;
+  }
+
+  async initGroupConfig(groupId: number) {
+    return this.groupConfigRepository.save({ groupId });
+  }
+
+  async updateGroupConfig({
+    groupId,
+    ...groupConfigData
+  }: UpdateGroupConfigInput) {
+    const group = await this.groupRepository.findOneOrFail({
+      where: { id: groupId },
+      relations: ['config'],
+    });
+    const newConfig = { ...group.config, ...groupConfigData };
+    if (
+      newConfig.decisionMakingModel === DecisionMakingModel.Consent &&
+      newConfig.votingTimeLimit === 0
+    ) {
+      throw new Error(
+        'Voting time limit is required for consent decision making model',
+      );
+    }
+
+    await this.groupConfigRepository.update(group.config.id, groupConfigData);
+    return { group };
+  }
+
+  async getGroupMemberRequest(
+    where: FindOptionsWhere<GroupMemberRequest>,
+    relations?: string[],
+  ) {
+    return this.groupMemberRequestRepository.findOne({
+      relations,
+      where,
+    });
+  }
+
+  async getGroupMemberRequests(groupId: number) {
+    return this.groupMemberRequestRepository.find({
+      where: { status: GroupMemberRequestStatus.Pending, groupId },
+      order: { createdAt: 'DESC' },
+    });
+  }
+
+  async getGroupMemberRequestCountBatch(groupIds: number[]) {
+    const groups = (await this.groupRepository
+      .createQueryBuilder('group')
+      .leftJoinAndSelect('group.memberRequests', 'memberRequest')
+      .loadRelationCountAndMap(
+        'group.memberRequestCount',
+        'group.memberRequests',
+        'memberRequest',
+        (qb) =>
+          qb.andWhere('memberRequest.status = :status', {
+            status: GroupMemberRequestStatus.Pending,
+          }),
+      )
+      .select(['group.id'])
+      .whereInIds(groupIds)
+      .getMany()) as GroupWithMemberRequestCount[];
+
+    return groupIds.map((id) => {
+      const group = groups.find((group: Group) => group.id === id);
+      if (!group) {
+        return new Error(`Could not load member request count: ${id}`);
+      }
+      return group.memberRequestCount;
+    });
+  }
+
+  async createGroupMemberRequest(groupId: number, userId: number) {
+    const groupMemberRequest = await this.groupMemberRequestRepository.save({
+      groupId,
+      userId,
+    });
+    return { groupMemberRequest };
+  }
+
+  async approveGroupMemberRequest(id: number) {
+    const memberRequest = await this.updateGroupMemberRequest(id, {
+      status: GroupMemberRequestStatus.Approved,
+    });
+    const groupMember = await this.createGroupMember(
+      memberRequest.groupId,
+      memberRequest.userId,
+    );
+    return { groupMember };
+  }
+
+  async denyGroupMemberRequest(id: number) {
+    await this.updateGroupMemberRequest(id, {
+      status: GroupMemberRequestStatus.Denied,
+    });
+    return true;
+  }
+
+  async updateGroupMemberRequest(
+    id: number,
+    requestData: Partial<GroupMemberRequest>,
+  ) {
+    await this.groupMemberRequestRepository.update(id, requestData);
+    return this.groupMemberRequestRepository.findOneOrFail({ where: { id } });
+  }
+
+  async cancelGroupMemberRequest(id: number) {
+    await this.deleteGroupMemberRequest({ id });
+    return true;
+  }
+
+  async deleteGroupMemberRequest(where: FindOptionsWhere<GroupMemberRequest>) {
+    await this.groupMemberRequestRepository.delete(where);
+    return true;
   }
 }
