+++ src/groups/models/approve-group-member-request.payload.ts
@@ -1,5 +1,5 @@
 import { Field, ObjectType } from '@nestjs/graphql';
-import { User } from '../../../users/models/user.model';
+import { User } from '../../users/models/user.model';
 
 @ObjectType()
 export class ApproveGroupMemberRequestPayload {
