+++ src/posts/posts.service.ts
@@ -5,7 +5,7 @@ import { FindOptionsWhere, In, Repository } from 'typeorm';
 import { DEFAULT_PAGE_SIZE } from '../common/common.constants';
 import { sanitizeText } from '../common/common.utils';
 import { IsLikedByMeKey } from '../dataloader/dataloader.types';
-import { GroupPrivacy } from '../groups/group-configs/group-configs.constants';
+import { GroupPrivacy } from '../groups/groups.constants';
 import { deleteImageFile, saveImage } from '../images/image.utils';
 import { Image } from '../images/models/image.model';
 import { Like } from '../likes/models/like.model';
