+++ src/shield/rules/comment.rules.ts
@@ -1,9 +1,9 @@
 import { rule } from 'graphql-shield';
 import { Comment } from '../../comments/models/comment.model';
 import { UpdateCommentInput } from '../../comments/models/update-comment.input';
-import { Context } from '../../context/context.types';
-import { GroupPrivacy } from '../../groups/group-configs/group-configs.constants';
 import { UNAUTHORIZED } from '../../common/common.constants';
+import { Context } from '../../context/context.types';
+import { GroupPrivacy } from '../../groups/groups.constants';
 
 export const isOwnComment = rule({ cache: 'strict' })(async (
   _parent,
