+++ src/shield/rules/event.rules.ts
@@ -2,7 +2,7 @@ import { rule } from 'graphql-shield';
 import { FeedItemsConnection } from '../../common/models/feed-items.connection';
 import { Context } from '../../context/context.types';
 import { Event } from '../../events/models/event.model';
-import { GroupPrivacy } from '../../groups/group-configs/group-configs.constants';
+import { GroupPrivacy } from '../../groups/groups.constants';
 import { Image } from '../../images/models/image.model';
 import { Post } from '../../posts/models/post.model';
 import { PostsConnection } from '../../posts/models/posts.connection';
