+++ src/users/models/user.model.ts
@@ -13,7 +13,7 @@ import {
 } from 'typeorm';
 import { Comment } from '../../comments/models/comment.model';
 import { EventAttendee } from '../../events/event-attendees/models/event-attendee.model';
-import { GroupMemberRequest } from '../../groups/group-member-requests/models/group-member-request.model';
+import { GroupMemberRequest } from '../../groups/models/group-member-request.model';
 import { GroupRole } from '../../groups/group-roles/models/group-role.model';
 import { Group } from '../../groups/models/group.model';
 import { Image } from '../../images/models/image.model';
