+++ src/users/users.service.ts
@@ -5,8 +5,8 @@ import { FileUpload } from 'graphql-upload-ts';
 import { FindOptionsWhere, In, Repository } from 'typeorm';
 import { logTime, paginate, sanitizeText } from '../common/common.utils';
 import { IsFollowedByMeKey } from '../dataloader/dataloader.types';
-import { GroupPrivacy } from '../groups/group-configs/group-configs.constants';
 import { GroupPermissionsMap } from '../groups/group-roles/models/group-permissions.type';
+import { GroupPrivacy } from '../groups/groups.constants';
 import { ImageTypes } from '../images/image.constants';
 import {
   deleteImageFile,
